Core/Loot Fix currency looting + switch min/maxCount to uint32

+++ src/server/game/Loot/LootMgr.h
@@ -43,13 +43,13 @@ struct TC_GAME_API LootStoreItem
     uint16  lootmode;
     bool    needs_quest;                                    // quest drop (quest is required for item to drop)
     uint8   groupid;
-    uint8   mincount;                                       // mincount for drop items
-    uint8   maxcount;                                       // max drop count for the item mincount or Ref multiplicator
+    uint32  mincount;                                       // mincount for drop items
+    uint32  maxcount;                                       // max drop count for the item mincount or Ref multiplicator
     ConditionContainer conditions;                               // additional loot condition
 
     // Constructor
     // displayid is filled in IsValid() which must be called after
-    LootStoreItem(uint32 _itemid, uint8 _type, uint32 _reference, float _chance, bool _needs_quest, uint16 _lootmode, uint8 _groupid, uint8 _mincount, uint8 _maxcount)
+    LootStoreItem(uint32 _itemid, uint8 _type, uint32 _reference, float _chance, bool _needs_quest, uint16 _lootmode, uint8 _groupid, uint32 _mincount, uint32 _maxcount)
         : itemid(_itemid), type(_type), reference(_reference), chance(_chance), lootmode(_lootmode),
         needs_quest(_needs_quest), groupid(_groupid), mincount(_mincount), maxcount(_maxcount)
          { }
